plot acoustic directivity using matlab code
The matlab inner function "polar" is not convinent for the arbitrary acoustic direcivity figure plotting due to that the size and the angle range (only allowing 360 degree) can't be revised. So here, the funciton "PlotDirecivity.m" is developed to plot the acoustic directivity more flexibly. Owing to individual habit, so you can revise it up to yourself.